Allows admission webhook to run on juju 2.9
With the introduction of juju 2.9 labels have changed to follow a more
Kubernetes approach by using standardised app labels and domain specific
Juju labels.

This changes the mutating webhook so that it now supports ignoring
objects that have these new labels avoiding a chicken and egg scenario.

The labels to support are:
- operator.juju.is/name
- app.kubernetes.io/name